# CTHz/openapi.yaml
openapi: 3.0.3
info:
  title: CTHz System 300 Device Management API (Internal)
  version: 0.1.0
  description: |
    Internal-only API specification for the System 300 Device Management (DM) service.

    Notes
    - This spec includes Spurtree-owned Device Management APIs and references CTHz-owned Sensor APIs (proxied by DM).
    - Do not expose Swagger/Redoc in production. Keep this file internal.
servers:
  - url: https://localhost:8443
    description: Local development (TLS)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: validation_error
        message:
          type: string
          example: Invalid request payload
        details:
          type: object
        requestId:
          type: string
          example: 9b2f9d62-4e1e-4a9b-95f1-3b9e0a9f3c1a
      required: [code, message]
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, monitor]
        is_active:
          type: boolean
    Device:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        device_id: { type: string }
        serial_number: { type: string }
        ip_address: { type: string }
        status: { type: string, example: online }
    Station:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        device_ids: { type: array, items: { type: string } }
    Policy:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        is_system_policy: { type: boolean }
        is_active: { type: boolean }
        conditions: { type: object }
        actions: { type: array, items: { type: object } }
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: User login authentication
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: JWT tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  refresh_token: { type: string }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /api/auth/setup:
    post:
      tags: [auth]
      summary: First-time setup (create Owner account)
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      x-rbac: [owner]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '201': { description: Owner created, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '409': { description: Already initialized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /api/auth/users:
    get:
      tags: [auth]
      summary: List all users (Admin only)
      x-owner: Spurtree
      x-priority: Medium
      x-phase: MVP
      x-rbac: [owner, admin]
      responses:
        '200': { description: Users, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/User' } } } } }
    post:
      tags: [auth]
      summary: Create new user (Admin only)
      x-owner: Spurtree
      x-priority: Medium
      x-phase: MVP
      x-rbac: [owner, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
  /api/auth/users/{id}:
    get:
      tags: [auth]
      summary: Get user details
      x-owner: Spurtree
      x-priority: Medium
      x-phase: MVP
      parameters:
        - in: path
          name: id
          schema: { type: string }
          required: true
      responses:
        '200': { description: User, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '404': { description: Not found }
    put:
      tags: [auth]
      summary: Update user
      x-owner: Spurtree
      x-priority: Medium
      x-phase: MVP
      parameters:
        - in: path
          name: id
          schema: { type: string }
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/User' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
    delete:
      tags: [auth]
      summary: Delete user
      x-owner: Spurtree
      x-priority: Medium
      x-phase: MVP
      parameters:
        - in: path
          name: id
          schema: { type: string }
          required: true
      responses:
        '204': { description: Deleted }
  /api/auth/users/{id}/role:
    put:
      tags: [auth]
      summary: Change user role
      x-owner: Spurtree
      x-priority: Medium
      x-phase: MVP
      x-rbac: [owner, admin]
      parameters:
        - in: path
          name: id
          schema: { type: string }
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [owner, admin, monitor]
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
...
  /api/devices:
    get:
      tags: [devices]
      summary: List managed devices
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      responses:
        '200': { description: Devices, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Device' } } } } }
    post:
      tags: [devices]
      summary: Add device manually
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Device' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Device' } } } }
...
  /api/stations:
    get:
      tags: [stations]
      summary: List all stations
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      responses:
        '200': { description: Stations, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Station' } } } } }
    post:
      tags: [stations]
      summary: Create new station
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Station' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Station' } } } }
...
  /api/policies:
    get:
      tags: [policies]
      summary: List all policies
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      responses:
        '200': { description: Policies, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Policy' } } } } }
    post:
      tags: [policies]
      summary: Create new policy
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Policy' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Policy' } } } }
...